I created the private reporter.
The vast majority of my flakes (all but one) are on github and I use gh to report issues.
But for the one that is private I can not use gh.
This just creates an issue branch which is exactly the same as main
except that it has one additional commit
The commit is empty and the message is based off arguments.